llamaServer/requirements.txt
websockets
gpt4all
python-dotenv